first iteration for implementation of db whitelistcheck
rudimentary check when player enters the game. if theplayer has any
special state attached to him, the server tells the player wether he is
whitelisted or not. there is no possibility to change the set permission
other than to kick the player and force him to request the permissions
again